requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blessed==1.20.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
django-picklefield==3.2
django-q2==1.6.2
groq==0.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
requests==2.32.3
setuptools==74.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
